created ICC, null model, and uncond growth model macro
This allowed me to create the google sheet here summarizing the issues with convergence and remaining outliers that need to be addressed (even after >|3|SD were removed in the first pass) in order to create valid trajectories for the repeated variables.

I also added some code to the top to eliminate observations on the input datasets that don't have subject IDs.